ai-services/llm-service/requirements.txt
flask==3.0.0
torch==2.1.1
transformers==4.36.0
peft==0.7.1
bitsandbytes==0.41.3
accelerate==0.25.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
lark-parser==0.12.0
